+++ AISKU/Tests/Unit/src/applicationinsights.e2e.tests.ts
@@ -8,7 +8,7 @@ import { IPropTelemetryContext } from '@microsoft/applicationinsights-properties
 import { createAsyncResolvedPromise } from '@nevware21/ts-async';
 import { CONFIG_ENDPOINT_URL } from '../../../src/InternalConstants';
 import { OfflineChannel } from '@microsoft/applicationinsights-offlinechannel-js';
-import { IStackFrame } from '@microsoft/applicationinsights-common/src/Interfaces/Contracts/IStackFrame';
+import { IStackFrame } from '@microsoft/applicationinsights-core-js';
 import { utcNow } from '@nevware21/ts-utils';
 
 function _checkExpectedFrame(expectedFrame: IStackFrame, actualFrame: IStackFrame, index: number) {