+++ shared/AppInsightsCore/Tests/Unit/src/AppInsightsCoreSize.Tests.ts
@@ -51,10 +51,10 @@ function _checkSize(checkType: string, maxSize: number, size: number, isNightly:
 }
 
 export class AppInsightsCoreSizeCheck extends AITestClass {
-    private readonly MAX_RAW_SIZE = 96;
-    private readonly MAX_BUNDLE_SIZE = 96;
-    private readonly MAX_RAW_DEFLATE_SIZE = 39;
-    private readonly MAX_BUNDLE_DEFLATE_SIZE = 39;
+    private readonly MAX_RAW_SIZE = 124;
+    private readonly MAX_BUNDLE_SIZE = 124;
+    private readonly MAX_RAW_DEFLATE_SIZE = 49;
+    private readonly MAX_BUNDLE_DEFLATE_SIZE = 49;
     private readonly rawFilePath = "../dist/es5/applicationinsights-core-js.min.js";
     private readonly prodFilePath = "../browser/es5/applicationinsights-core-js.min.js";