+++ shared/AppInsightsCore/Tests/Unit/src/Common/appinsights-common.tests.ts
@@ -5,7 +5,7 @@ import { ConnectionStringParserTests } from "./ConnectionStringParser.tests";
 import { SeverityLevelTests } from "./SeverityLevel.tests";
 import { RequestHeadersTests } from "./RequestHeaders.tests";
 import { ThrottleMgrTest } from "./ThrottleMgr.tests";
-import { GlobalTestHooks } from "./GlobalTestHooks.Test";
+import { GlobalTestHooks } from "../GlobalTestHooks.Test";
 import { W3CTraceStateModesTests } from "./W3CTraceStateModes.tests";
 
 export function runTests() {