+++ shared/AppInsightsCore/src/Common/HelperFuncs.ts
@@ -1,8 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { IPlugin, arrForEach, isString, isTimeSpan } from "@microsoft/applicationinsights-core-js";
-import { mathFloor, mathRound } from "@nevware21/ts-utils";
+import { arrForEach, isString, mathFloor, mathRound } from "@nevware21/ts-utils";
+import { IPlugin } from "../JavaScriptSDK.Interfaces/ITelemetryPlugin";
+import { isTimeSpan } from "../OpenTelemetry/helpers/timeHelpers";
 
 const strEmpty = "";