+++ shared/AppInsightsCore/src/Common/Interfaces/Context/ISample.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { ITelemetryItem } from "@microsoft/applicationinsights-core-js";
+import { ITelemetryItem } from "../../../JavaScriptSDK.Interfaces/ITelemetryItem";
 
 export interface ISample {
     /**