+++ shared/AppInsightsCore/src/Common/Interfaces/IAppInsights.ts
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { ICookieMgr, ITelemetryItem } from "@microsoft/applicationinsights-core-js";
+import { ICookieMgr } from "../../JavaScriptSDK.Interfaces/ICookieMgr";
+import { ITelemetryItem } from "../../JavaScriptSDK.Interfaces/ITelemetryItem";
 import { IEventTelemetry } from "./IEventTelemetry";
 import { IAutoExceptionTelemetry, IExceptionTelemetry } from "./IExceptionTelemetry";
 import { IMetricTelemetry } from "./IMetricTelemetry";